MINOR: http-ana: Add a HTTP_MSGF flag to state the Expect header was checked
HTTP_MSGF_EXPECT_CHECKED is now set on the request message to know the
"Expect: " header was already handled, if any. The flag is set from the
moment we try to handle the header to send a "100-continue" response,
whether it was found or not.

This way, when we are waiting for the request payload, thanks to this flag,
we only try to handle "Expect: " header only once. Before it was performed
by changing the message state from BODY to DATA. But this has some side
effects and it is no accurate. So, it is better to rely on a flag to do so.

(cherry picked from commit ffcffa8e93bcbf376f13b1f6e568ab093ea068bf)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit ada2f085a4ec107a1d7e5436b14e8b271e54ae87)
[cf: hmsg_show_flags() function does not exist on 2.6]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>