MINOR: quic: Add a trace for packet with an ACK frame
As the ACK frames are not added to the packet list of ack-eliciting frames,
it could not be traced. But there is a flag to identify such packet.
Let's use it to add this information to the traces of TX packets.

Must be backported to 2.6 and 2.7.

(cherry picked from commit 6fd2576d5ebbe90039cc918d01d0f62970f7cd67)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>