MINOR: activity: add a line reporting the average CPU usage to "show activity"
It was missing from the output but is sometimes convenient to observe
and understand how incoming connections are distributed. The CPU usage
is reported as the instant measurement of 100-idle_pct for each thread,
and the average value is shown for the aggregated value.

This could be backported as it's helpful in certain troublehsooting
sessions.

(cherry picked from commit 28f2a590f63a2a051fe8cfc64b74dfa06c9e7f32)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>