Revert "BUG/MEDIUM: stconn: Don't rearm the read expiration date if EOI was reached"
This reverts commit 5eb0053ce16d1b9242eece90ae351afcdc7592e2.

This commit revealed hidden bugs leading to session freeze or to spinning
loop in process_stream. Some of these bugs cannot be reproduced easily. The
2.8-DEV is not affected. This patch must only be reverted stable versions.
For now, only 2.7, 2.6 and 2.5. Instead of introducing instablility trying to
blindly fix these bugs, it is safer to revert this patch. At least for
now. The only drawback is to have some server timeouts reported instead of
client timeouts.

This patch shoudld fix the issues #2111, #2090 and #2088. It must be
backported where the above commit was backported too.